Added transform function

+++ src/java/com/pulsar/ui/engine/Vector.java
@@ -8,6 +8,6 @@ public interface Vector {
 
 	Color getFillColor();
 
-	Vector transform(long distance, double angle, long radius, long screenSize, int screenWidth, long minSize);
+	Vector transform(long distance, double angle, long radius, long screenSize, int screenWidth, int minSize);
 
 }
